Revert "Feature/tfp 3040 (#447)"
This reverts commit cfcb3b2a848b571bc7b8dc1042a792c23d9b89c5.
