+++ domenetjenester/iay/src/main/java/no/nav/foreldrepenger/abakus/domene/iay/InntektspostBuilder.java
@@ -7,8 +7,6 @@ import no.nav.abakus.iaygrunnlag.kodeverk.InntektspostType;
 import no.nav.abakus.iaygrunnlag.kodeverk.SkatteOgAvgiftsregelType;
 import no.nav.abakus.iaygrunnlag.kodeverk.UtbetaltYtelseType;
 import no.nav.foreldrepenger.abakus.typer.Beløp;
-import no.nav.foreldrepenger.abakus.typer.OrgNummer;
-import no.nav.foreldrepenger.abakus.typer.OrganisasjonsNummerValidator;
 
 public class InntektspostBuilder {
     private Inntektspost inntektspost;
@@ -38,15 +36,6 @@ public class InntektspostBuilder {
 
     public InntektspostBuilder medBeløp(BigDecimal verdi) {
         this.inntektspost.setBeløp(new Beløp(verdi));
-        return this;
-    }
-
-    public InntektspostBuilder medOpprinneligUtbetalerOrgnr(String opprinneligUtbetalerOrgnr) {
-        if (!OrganisasjonsNummerValidator.erGyldig(opprinneligUtbetalerOrgnr)) {
-            throw new IllegalStateException("Ikke gyldig orgnr for opprinnelig utbetaler");
-        }
-        OrgNummer orgNummer = new OrgNummer(opprinneligUtbetalerOrgnr);
-        this.inntektspost.setOpprinneligUtbetalerOrgnr(orgNummer);
         return this;
     }
