+++ domenetjenester/iay/src/main/java/no/nav/foreldrepenger/abakus/registerdata/IAYRegisterInnhentingFellesTjenesteImpl.java
@@ -1,23 +1,36 @@
 package no.nav.foreldrepenger.abakus.registerdata;
 
 import java.math.BigDecimal;
+import java.time.LocalDate;
+import java.time.YearMonth;
 import java.util.Collections;
+import java.util.Comparator;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
+import java.util.stream.Collectors;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.threeten.extra.Interval;
 
 import no.nav.abakus.iaygrunnlag.kodeverk.ArbeidType;
 import no.nav.abakus.iaygrunnlag.kodeverk.InntektskildeType;
 import no.nav.abakus.iaygrunnlag.kodeverk.InntektspostType;
+import no.nav.abakus.iaygrunnlag.kodeverk.SkatteOgAvgiftsregelType;
+import no.nav.abakus.iaygrunnlag.kodeverk.UtbetaltNæringsYtelseType;
+import no.nav.abakus.iaygrunnlag.kodeverk.UtbetaltPensjonTrygdType;
+import no.nav.abakus.iaygrunnlag.kodeverk.UtbetaltYtelseFraOffentligeType;
+import no.nav.abakus.iaygrunnlag.kodeverk.UtbetaltYtelseType;
 import no.nav.foreldrepenger.abakus.domene.iay.AktivitetsAvtaleBuilder;
 import no.nav.foreldrepenger.abakus.domene.iay.Arbeidsgiver;
 import no.nav.foreldrepenger.abakus.domene.iay.InntektArbeidYtelseAggregatBuilder;
 import no.nav.foreldrepenger.abakus.domene.iay.InntektArbeidYtelseGrunnlag;
 import no.nav.foreldrepenger.abakus.domene.iay.InntektArbeidYtelseGrunnlagBuilder;
+import no.nav.foreldrepenger.abakus.domene.iay.InntektBuilder;
+import no.nav.foreldrepenger.abakus.domene.iay.InntektspostBuilder;
 import no.nav.foreldrepenger.abakus.domene.iay.Opptjeningsnøkkel;
 import no.nav.foreldrepenger.abakus.domene.iay.YrkesaktivitetBuilder;
 import no.nav.foreldrepenger.abakus.domene.iay.arbeidsforhold.ArbeidsforholdInformasjon;
@@ -31,23 +44,29 @@ import no.nav.foreldrepenger.abakus.registerdata.arbeidsforhold.Organisasjon;
 import no.nav.foreldrepenger.abakus.registerdata.arbeidsforhold.Person;
 import no.nav.foreldrepenger.abakus.registerdata.arbeidsgiver.virksomhet.VirksomhetTjeneste;
 import no.nav.foreldrepenger.abakus.registerdata.inntekt.komponenten.FrilansArbeidsforhold;
+import no.nav.foreldrepenger.abakus.registerdata.inntekt.komponenten.InntektsInformasjon;
+import no.nav.foreldrepenger.abakus.registerdata.inntekt.komponenten.Månedsinntekt;
 import no.nav.foreldrepenger.abakus.registerdata.inntekt.sigrun.SigrunTjeneste;
 import no.nav.foreldrepenger.abakus.registerdata.tjeneste.RegisterdataElement;
 import no.nav.foreldrepenger.abakus.typer.AktørId;
 import no.nav.foreldrepenger.abakus.typer.EksternArbeidsforholdRef;
 import no.nav.foreldrepenger.abakus.typer.InternArbeidsforholdRef;
+import no.nav.foreldrepenger.abakus.typer.OrganisasjonsNummerValidator;
+import no.nav.foreldrepenger.abakus.typer.PersonIdent;
 import no.nav.foreldrepenger.abakus.vedtak.domene.VedtakYtelseRepository;
+import no.nav.vedtak.felles.integrasjon.aktør.klient.AktørConsumer;
 
 public abstract class IAYRegisterInnhentingFellesTjenesteImpl implements IAYRegisterInnhentingTjeneste {
 
     public static final Map<RegisterdataElement, InntektskildeType> ELEMENT_TIL_INNTEKTS_KILDE_MAP = Map.of(RegisterdataElement.INNTEKT_PENSJONSGIVENDE, InntektskildeType.INNTEKT_OPPTJENING, RegisterdataElement.INNTEKT_BEREGNINGSGRUNNLAG, InntektskildeType.INNTEKT_BEREGNING, RegisterdataElement.INNTEKT_SAMMENLIGNINGSGRUNNLAG, InntektskildeType.INNTEKT_SAMMENLIGNING);
+    private static final Logger LOGGER = LoggerFactory.getLogger(IAYRegisterInnhentingFellesTjenesteImpl.class);
     protected InntektArbeidYtelseTjeneste inntektArbeidYtelseTjeneste;
     private VirksomhetTjeneste virksomhetTjeneste;
     private YtelseRegisterInnhenting ytelseRegisterInnhenting;
     private InnhentingSamletTjeneste innhentingSamletTjeneste;
     private ByggYrkesaktiviteterTjeneste byggYrkesaktiviteterTjeneste;
+    private AktørConsumer aktørConsumer;
     private SigrunTjeneste sigrunTjeneste;
-    private MapInntektFraDtoTilDomene mapInntektFraDtoTilDomene;
 
     protected IAYRegisterInnhentingFellesTjenesteImpl() {
     }
@@ -55,16 +74,35 @@ public abstract class IAYRegisterInnhentingFellesTjenesteImpl implements IAYRegi
     protected IAYRegisterInnhentingFellesTjenesteImpl(InntektArbeidYtelseTjeneste inntektArbeidYtelseTjeneste,
                                                       VirksomhetTjeneste virksomhetTjeneste,
                                                       InnhentingSamletTjeneste innhentingSamletTjeneste,
-                                                      SigrunTjeneste sigrunTjeneste,
-                                                      VedtakYtelseRepository vedtakYtelseRepository,
-                                                      MapInntektFraDtoTilDomene mapInntektFraDtoTilDomene) {
+                                                      AktørConsumer aktørConsumer,
+                                                      SigrunTjeneste sigrunTjeneste, VedtakYtelseRepository vedtakYtelseRepository) {
         this.inntektArbeidYtelseTjeneste = inntektArbeidYtelseTjeneste;
         this.virksomhetTjeneste = virksomhetTjeneste;
         this.innhentingSamletTjeneste = innhentingSamletTjeneste;
+        this.aktørConsumer = aktørConsumer;
         this.sigrunTjeneste = sigrunTjeneste;
         this.ytelseRegisterInnhenting = new YtelseRegisterInnhenting(innhentingSamletTjeneste, vedtakYtelseRepository);
         this.byggYrkesaktiviteterTjeneste = new ByggYrkesaktiviteterTjeneste();
-        this.mapInntektFraDtoTilDomene = mapInntektFraDtoTilDomene;
+    }
+
+    private void innhentNæringsOpplysninger(Kobling kobling, InntektArbeidYtelseAggregatBuilder inntektArbeidYtelseAggregatBuilder) {
+        var map = sigrunTjeneste.beregnetSkatt(kobling.getAktørId());
+        var aktørInntektBuilder = inntektArbeidYtelseAggregatBuilder.getAktørInntektBuilder(kobling.getAktørId());
+
+        var inntektBuilder = aktørInntektBuilder.getInntektBuilder(InntektskildeType.SIGRUN, null);
+
+        for (var entry : map.entrySet()) {
+            for (Map.Entry<InntektspostType, BigDecimal> type : entry.getValue().entrySet()) {
+                InntektspostBuilder inntektspostBuilder = inntektBuilder.getInntektspostBuilder();
+                inntektspostBuilder
+                    .medInntektspostType(type.getKey())
+                    .medBeløp(type.getValue())
+                    .medPeriode(entry.getKey().getFomDato(), entry.getKey().getTomDato());
+                inntektBuilder.leggTilInntektspost(inntektspostBuilder);
+            }
+        }
+        aktørInntektBuilder.leggTilInntekt(inntektBuilder);
+        inntektArbeidYtelseAggregatBuilder.leggTilAktørInntekt(aktørInntektBuilder);
     }
 
     @Override
@@ -95,11 +133,6 @@ public abstract class IAYRegisterInnhentingFellesTjenesteImpl implements IAYRegi
         return grunnlagBuilder;
     }
 
-    private void innhentNæringsOpplysninger(Kobling kobling, InntektArbeidYtelseAggregatBuilder inntektArbeidYtelseAggregatBuilder) {
-        Map<IntervallEntitet, Map<InntektspostType, BigDecimal>> map = sigrunTjeneste.beregnetSkatt(kobling.getAktørId());
-        mapInntektFraDtoTilDomene.mapFraSigrun(kobling.getAktørId(), map, inntektArbeidYtelseAggregatBuilder);
-    }
-
     private void innhentYtelser(Kobling kobling, InntektArbeidYtelseAggregatBuilder builder) {
         ytelseRegisterInnhenting.byggYtelser(kobling, kobling.getAktørId(),
             kobling.getOpplysningsperiode().tilIntervall(),
@@ -109,6 +142,31 @@ public abstract class IAYRegisterInnhentingFellesTjenesteImpl implements IAYRegi
 
     private Set<ArbeidsforholdIdentifikator> innhentArbeidsforhold(Kobling kobling, InntektArbeidYtelseAggregatBuilder builder, Set<RegisterdataElement> informasjonsElementer) {
         return byggOpptjeningOpplysningene(kobling, kobling.getAktørId(), kobling.getOpplysningsperiode().tilIntervall(), builder, informasjonsElementer);
+    }
+
+    private void leggTilInntekter(AktørId aktørId, InntektArbeidYtelseAggregatBuilder builder, InntektsInformasjon inntektsInformasjon) {
+        var aktørInntektBuilder = builder.getAktørInntektBuilder(aktørId);
+        InntektskildeType kilde = inntektsInformasjon.getKilde();
+        aktørInntektBuilder.fjernInntekterFraKilde(kilde);
+
+        inntektsInformasjon.getMånedsinntekterGruppertPåArbeidsgiver()
+            .forEach((identifikator, inntektOgRegelListe) -> leggTilInntekterPåArbeidsforhold(builder, aktørInntektBuilder, inntektOgRegelListe, identifikator,
+                kilde));
+
+        List<Månedsinntekt> ytelsesTrygdEllerPensjonInntekt = inntektsInformasjon.getYtelsesTrygdEllerPensjonInntektSummert();
+        if (!ytelsesTrygdEllerPensjonInntekt.isEmpty()) {
+            leggTilYtelseInntekter(ytelsesTrygdEllerPensjonInntekt, builder, aktørId, kilde);
+        }
+    }
+
+    private void leggTilYtelseInntekter(List<Månedsinntekt> ytelsesTrygdEllerPensjonInntekt, InntektArbeidYtelseAggregatBuilder builder, AktørId aktørId,
+                                        InntektskildeType inntektOpptjening) {
+        var aktørInntektBuilder = builder.getAktørInntektBuilder(aktørId);
+        var inntektBuilderForYtelser = aktørInntektBuilder.getInntektBuilderForYtelser(inntektOpptjening);
+        ytelsesTrygdEllerPensjonInntekt.forEach(mi -> lagInntektsposterYtelse(mi, inntektBuilderForYtelser));
+
+        aktørInntektBuilder.leggTilInntekt(inntektBuilderForYtelser);
+        builder.leggTilAktørInntekt(aktørInntektBuilder);
     }
 
     private void oversettFrilanseArbeidsforhold(Kobling kobling,
@@ -142,6 +200,44 @@ public abstract class IAYRegisterInnhentingFellesTjenesteImpl implements IAYRegi
 
         aktørArbeidBuilder.leggTilYrkesaktivitet(yrkesaktivitetBuilder);
         builder.leggTilAktørArbeid(aktørArbeidBuilder);
+    }
+
+    private void leggTilInntekterPåArbeidsforhold(InntektArbeidYtelseAggregatBuilder builder,
+                                                  InntektArbeidYtelseAggregatBuilder.AktørInntektBuilder aktørInntektBuilder,
+                                                  Map<YearMonth, List<InntektsInformasjon.MånedsbeløpOgSkatteOgAvgiftsregel>> månedsinntekterGruppertPåArbeidsgiver,
+                                                  String arbeidsgiverIdentifikator, InntektskildeType inntektOpptjening) {
+
+        Arbeidsgiver arbeidsgiver;
+        if (OrganisasjonsNummerValidator.erGyldig(arbeidsgiverIdentifikator)) {
+            boolean orgledd = virksomhetTjeneste.sjekkOmVirksomhetErOrgledd(arbeidsgiverIdentifikator);
+            if (!orgledd) {
+                LocalDate hentedato = finnHentedatoForJuridisk(månedsinntekterGruppertPåArbeidsgiver.keySet());
+                arbeidsgiver = Arbeidsgiver.virksomhet(virksomhetTjeneste.hentOgLagreOrganisasjonMedHensynTilJuridisk(arbeidsgiverIdentifikator, hentedato));
+                aktørInntektBuilder.leggTilInntekt(byggInntekt(månedsinntekterGruppertPåArbeidsgiver, arbeidsgiver, aktørInntektBuilder, inntektOpptjening));
+                builder.leggTilAktørInntekt(aktørInntektBuilder);
+            } else {
+                LOGGER.info("Inntekter rapportert på orglegg({}), blir IKKE lagret", arbeidsgiverIdentifikator);
+            }
+        } else {
+            if (PersonIdent.erGyldigFnr(arbeidsgiverIdentifikator)) {
+                Optional<String> arbeidsgiverOpt = aktørConsumer.hentAktørIdForPersonIdent(arbeidsgiverIdentifikator);
+                if (arbeidsgiverOpt.isEmpty()) {
+                    throw InnhentingFeil.FACTORY.finnerIkkeAktørIdForArbeidsgiverSomErPrivatperson().toException();
+                }
+                arbeidsgiver = Arbeidsgiver.person(new AktørId(arbeidsgiverOpt.get()));
+            } else {
+                LOGGER.info("Arbeidsgiveridentifikator: {}", arbeidsgiverIdentifikator);
+                arbeidsgiver = Arbeidsgiver.person(new AktørId(arbeidsgiverIdentifikator));
+            }
+            aktørInntektBuilder.leggTilInntekt(byggInntekt(månedsinntekterGruppertPåArbeidsgiver, arbeidsgiver, aktørInntektBuilder, inntektOpptjening));
+            builder.leggTilAktørInntekt(aktørInntektBuilder);
+        }
+    }
+
+    private LocalDate finnHentedatoForJuridisk(Set<YearMonth> inntekterForMåneder) {
+        return inntekterForMåneder.stream()
+            .map(m -> LocalDate.of(m.getYear(), m.getMonth(), 1))
+            .max(Comparator.naturalOrder()).orElse(LocalDate.now());
     }
 
     private Set<ArbeidsforholdIdentifikator> byggOpptjeningOpplysningene(Kobling kobling, AktørId aktørId, Interval opplysningsPeriode,
@@ -181,7 +277,7 @@ public abstract class IAYRegisterInnhentingFellesTjenesteImpl implements IAYRegi
         var inntektsInformasjon = innhentingSamletTjeneste.getInntektsInformasjon(aktørId, kobling.getOpplysningsperiode().tilIntervall(), inntektsKilde);
 
         if (informasjonsElementer.contains(registerdataElement)) {
-            mapInntektFraDtoTilDomene.mapFraInntektskomponent(aktørId, builder, inntektsInformasjon);
+            leggTilInntekter(aktørId, builder, inntektsInformasjon);
         }
 
         if (inntektsKilde.equals(InntektskildeType.INNTEKT_OPPTJENING) && informasjonsElementer.contains(RegisterdataElement.ARBEIDSFORHOLD)) {
@@ -244,5 +340,77 @@ public abstract class IAYRegisterInnhentingFellesTjenesteImpl implements IAYRegi
             periode = IntervallEntitet.fraOgMedTilOgMed(frilansArbeidsforhold.getFom(), frilansArbeidsforhold.getTom());
         }
         return yrkesaktivitetBuilder.getAktivitetsAvtaleBuilder(periode, true);
+    }
+
+    private InntektBuilder byggInntekt(Map<YearMonth, List<InntektsInformasjon.MånedsbeløpOgSkatteOgAvgiftsregel>> inntekter,
+                                       Arbeidsgiver arbeidsgiver,
+                                       InntektArbeidYtelseAggregatBuilder.AktørInntektBuilder aktørInntektBuilder,
+                                       InntektskildeType inntektOpptjening) {
+
+        InntektBuilder inntektBuilder = aktørInntektBuilder.getInntektBuilder(inntektOpptjening, new Opptjeningsnøkkel(arbeidsgiver));
+
+        for (var måned : inntekter.keySet()) {
+            var månedsinnteker = inntekter.get(måned);
+            Map<String, Integer> antalInntekterForAvgiftsregel = månedsinnteker
+                .stream()
+                .filter(e -> e.getSkatteOgAvgiftsregelType() != null)
+                .collect(Collectors.groupingBy(
+                    InntektsInformasjon.MånedsbeløpOgSkatteOgAvgiftsregel::getSkatteOgAvgiftsregelType,
+                    Collectors.collectingAndThen(
+                        Collectors.mapping(InntektsInformasjon.MånedsbeløpOgSkatteOgAvgiftsregel::getBeløp, Collectors.toSet()),
+                        Set::size)));
+
+            Optional<String> valgtSkatteOgAvgiftsregel = Optional.empty();
+            BigDecimal beløpSum = månedsinnteker.stream().map(InntektsInformasjon.MånedsbeløpOgSkatteOgAvgiftsregel::getBeløp).reduce(BigDecimal.ZERO,
+                BigDecimal::add);
+            if (antalInntekterForAvgiftsregel.keySet().size() > 1) {
+                String skatteOgAvgiftsregler = antalInntekterForAvgiftsregel.keySet().stream().collect(Collectors.joining(", "));
+                // TODO Diamant velger her en random verdi.
+                valgtSkatteOgAvgiftsregel = Optional.of(antalInntekterForAvgiftsregel.keySet().iterator().next());
+                LOGGER.error("ArbeidsgiverEntitet orgnr {} har flere månedsinntekter for måned {} med forskjellige skatte -og avgiftsregler {}. Velger {}",
+                    arbeidsgiver.getIdentifikator(), måned, skatteOgAvgiftsregler, valgtSkatteOgAvgiftsregel);
+            } else if (antalInntekterForAvgiftsregel.keySet().size() == 1) {
+                valgtSkatteOgAvgiftsregel = Optional.of(antalInntekterForAvgiftsregel.keySet().iterator().next());
+            }
+
+            lagInntektsposter(måned, beløpSum, valgtSkatteOgAvgiftsregel, inntektBuilder);
+        }
+
+        return inntektBuilder
+            .medArbeidsgiver(arbeidsgiver);
+    }
+
+    private void lagInntektsposterYtelse(Månedsinntekt månedsinntekt, InntektBuilder inntektBuilder) {
+        inntektBuilder.leggTilInntektspost(inntektBuilder.getInntektspostBuilder()
+            .medBeløp(månedsinntekt.getBeløp())
+            .medPeriode(månedsinntekt.getMåned().atDay(1), månedsinntekt.getMåned().atEndOfMonth())
+            .medInntektspostType(InntektspostType.YTELSE)
+            .medYtelse(mapTilKodeliste(månedsinntekt)));
+    }
+
+    private void lagInntektsposter(YearMonth måned, BigDecimal sumInntektsbeløp, Optional<String> valgtSkatteOgAvgiftsregel,
+                                   InntektBuilder inntektBuilder) {
+
+        InntektspostBuilder inntektspostBuilder = inntektBuilder.getInntektspostBuilder();
+        inntektspostBuilder
+            .medBeløp(sumInntektsbeløp)
+            .medPeriode(måned.atDay(1), måned.atEndOfMonth())
+            .medInntektspostType(InntektspostType.LØNN);
+
+        if (valgtSkatteOgAvgiftsregel.isPresent()) {
+            var skatteOgAvgiftsregelType = SkatteOgAvgiftsregelType.finnForKodeverkEiersKode(valgtSkatteOgAvgiftsregel.get());
+            inntektspostBuilder.medSkatteOgAvgiftsregelType(skatteOgAvgiftsregelType);
+        }
+
+        inntektBuilder.leggTilInntektspost(inntektspostBuilder);
+    }
+
+    private UtbetaltYtelseType mapTilKodeliste(Månedsinntekt månedsinntekt) {
+        if (månedsinntekt.getPensjonKode() != null) {
+            return UtbetaltPensjonTrygdType.finnForKodeverkEiersKode(månedsinntekt.getPensjonKode());
+        } else if (månedsinntekt.getYtelseKode() != null) {
+            return UtbetaltYtelseFraOffentligeType.finnForKodeverkEiersKode(månedsinntekt.getYtelseKode());
+        }
+        return UtbetaltNæringsYtelseType.finnForKodeverkEiersKode(månedsinntekt.getNæringsinntektKode());
     }
 }
