+++ domenetjenester/iay/src/main/java/no/nav/foreldrepenger/abakus/registerdata/fagsakytelser/es/IAYRegisterInnhentingESTjenesteImpl.java
@@ -8,10 +8,10 @@ import no.nav.foreldrepenger.abakus.kobling.Kobling;
 import no.nav.foreldrepenger.abakus.kobling.kontroll.YtelseTypeRef;
 import no.nav.foreldrepenger.abakus.registerdata.IAYRegisterInnhentingFellesTjenesteImpl;
 import no.nav.foreldrepenger.abakus.registerdata.InnhentingSamletTjeneste;
-import no.nav.foreldrepenger.abakus.registerdata.MapInntektFraDtoTilDomene;
 import no.nav.foreldrepenger.abakus.registerdata.arbeidsgiver.virksomhet.VirksomhetTjeneste;
 import no.nav.foreldrepenger.abakus.registerdata.inntekt.sigrun.SigrunTjeneste;
 import no.nav.foreldrepenger.abakus.vedtak.domene.VedtakYtelseRepository;
+import no.nav.vedtak.felles.integrasjon.aktør.klient.AktørConsumer;
 
 @ApplicationScoped
 @YtelseTypeRef("ES")
@@ -25,13 +25,13 @@ public class IAYRegisterInnhentingESTjenesteImpl extends IAYRegisterInnhentingFe
     public IAYRegisterInnhentingESTjenesteImpl(InntektArbeidYtelseTjeneste inntektArbeidYtelseTjeneste,
                                                VirksomhetTjeneste virksomhetTjeneste,
                                                InnhentingSamletTjeneste innhentingSamletTjeneste,
-                                               SigrunTjeneste sigrunTjeneste,
-                                               VedtakYtelseRepository vedtakYtelseRepository,
-                                               MapInntektFraDtoTilDomene mapInntektFraDtoTilDomene) {
+                                               AktørConsumer aktørConsumer,
+                                               SigrunTjeneste sigrunTjeneste, VedtakYtelseRepository vedtakYtelseRepository) {
         super(inntektArbeidYtelseTjeneste,
             virksomhetTjeneste,
             innhentingSamletTjeneste,
-            sigrunTjeneste, vedtakYtelseRepository, mapInntektFraDtoTilDomene);
+            aktørConsumer,
+            sigrunTjeneste, vedtakYtelseRepository);
     }
 
     @Override
