+++ domenetjenester/iay/src/main/java/no/nav/foreldrepenger/abakus/registerdata/fagsakytelser/omp/OmsorgspengerIAYRegisterInnhentingTjenesteImpl.java
@@ -8,7 +8,6 @@ import no.nav.foreldrepenger.abakus.kobling.Kobling;
 import no.nav.foreldrepenger.abakus.kobling.kontroll.YtelseTypeRef;
 import no.nav.foreldrepenger.abakus.registerdata.IAYRegisterInnhentingFellesTjenesteImpl;
 import no.nav.foreldrepenger.abakus.registerdata.InnhentingSamletTjeneste;
-import no.nav.foreldrepenger.abakus.registerdata.MapInntektFraDtoTilDomene;
 import no.nav.foreldrepenger.abakus.registerdata.arbeidsgiver.virksomhet.VirksomhetTjeneste;
 import no.nav.foreldrepenger.abakus.registerdata.inntekt.sigrun.SigrunTjeneste;
 import no.nav.foreldrepenger.abakus.vedtak.domene.VedtakYtelseRepository;
@@ -29,12 +28,13 @@ public class OmsorgspengerIAYRegisterInnhentingTjenesteImpl extends IAYRegisterI
     public OmsorgspengerIAYRegisterInnhentingTjenesteImpl(InntektArbeidYtelseTjeneste inntektArbeidYtelseTjeneste,
                                                           VirksomhetTjeneste virksomhetTjeneste,
                                                           InnhentingSamletTjeneste innhentingSamletTjeneste,
-                                                          SigrunTjeneste sigrunTjeneste, VedtakYtelseRepository vedtakYtelseRepository,
-                                                          MapInntektFraDtoTilDomene mapInntektFraDtoTilDomene) {
+                                                          AktørConsumer aktørConsumer,
+                                                          SigrunTjeneste sigrunTjeneste, VedtakYtelseRepository vedtakYtelseRepository) {
         super(inntektArbeidYtelseTjeneste,
             virksomhetTjeneste,
             innhentingSamletTjeneste,
-            sigrunTjeneste, vedtakYtelseRepository, mapInntektFraDtoTilDomene);
+            aktørConsumer,
+            sigrunTjeneste, vedtakYtelseRepository);
     }
 
     @Override
